# apps/media/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WallpaperDB Media API",
    "version": "1.0.0",
    "description": "Delivers original and resized wallpapers and current Profile pictures."
  },
  "components": {
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["healthy", "unhealthy", "degraded", "error", "shutting_down"],
            "description": "Overall health status of the service"
          },
          "checks": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean"
            },
            "description": "Individual component health checks"
          },
          "timestamp": {
            "type": "string",
            "description": "ISO 8601 timestamp of the health check"
          },
          "error": {
            "type": "string",
            "description": "Error message if status is 'error'"
          },
          "totalDurationMs": {
            "type": "number",
            "description": "Duration of all checks in milliseconds"
          }
        },
        "required": ["status", "checks", "timestamp"],
        "additionalProperties": false
      },
      "ReadyResponse": {
        "type": "object",
        "properties": {
          "ready": {
            "type": "boolean",
            "description": "Whether the service is ready to accept traffic"
          },
          "timestamp": {
            "type": "string",
            "description": "ISO 8601 timestamp of the readiness check"
          },
          "reason": {
            "type": "string",
            "description": "Reason if service is not ready"
          }
        },
        "required": ["ready", "timestamp"],
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "URI identifying the problem type"
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary of the problem"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599,
            "description": "HTTP status code"
          },
          "detail": {
            "type": "string",
            "description": "Human-readable explanation specific to this occurrence of the problem"
          },
          "instance": {
            "type": "string",
            "description": "URI reference identifying the specific occurrence"
          }
        },
        "additionalProperties": true,
        "description": "RFC 7807 Problem Details response"
      },
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "URI identifying the problem type"
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary of the problem"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599,
            "description": "HTTP status code"
          },
          "detail": {
            "type": "string",
            "description": "Human-readable explanation specific to this occurrence of the problem"
          },
          "instance": {
            "type": "string",
            "description": "URI reference identifying the specific occurrence"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "description": "Field that failed validation"
                },
                "message": {
                  "type": "string",
                  "description": "Validation error message"
                },
                "code": {
                  "type": "string",
                  "description": "Error code for programmatic handling"
                }
              },
              "required": ["field", "message"],
              "additionalProperties": false
            },
            "description": "Array of validation errors"
          }
        },
        "additionalProperties": true,
        "description": "RFC 7807 Problem Details response"
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Dependency health",
        "tags": ["Health"],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service is degraded, unhealthy, or shutting down",
            "content": {
              "application/problem+json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "required": [
                        "type",
                        "title",
                        "status",
                        "timestamp",
                        "healthStatus",
                        "checks"
                      ],
                      "properties": {
                        "status": {
                          "type": "integer",
                          "enum": [503]
                        },
                        "timestamp": {
                          "type": "string"
                        },
                        "healthStatus": {
                          "type": "string",
                          "enum": ["degraded", "unhealthy", "shutting_down"]
                        },
                        "checks": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "boolean"
                          }
                        },
                        "totalDurationMs": {
                          "type": "number"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "summary": "Readiness",
        "tags": ["Health"],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadyResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service is not initialized or is shutting down",
            "content": {
              "application/problem+json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "required": ["type", "title", "status", "timestamp", "ready", "reason"],
                      "properties": {
                        "status": {
                          "type": "integer",
                          "enum": [503]
                        },
                        "timestamp": {
                          "type": "string"
                        },
                        "ready": {
                          "type": "boolean",
                          "enum": [false]
                        },
                        "reason": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/wallpapers/{id}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/profile-pictures/{pictureId}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "pictureId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3003",
      "description": "Local development server"
    }
  ]
}